+++ src/com/dabomstew/pkrandom/newnds/NDSRom.java
@@ -64,13 +64,9 @@ public class NDSRom {
         this.baseRom = new RandomAccessFile(filename, "r");
         this.romOpen = true;
         // TMP folder?
-        String rawFilename = new File(filename).getName();
-        String dataFolder = "tmp_" + rawFilename.substring(0, rawFilename.lastIndexOf('.'));
-        // remove nonsensical chars
-        dataFolder = dataFolder.replaceAll("[^A-Za-z0-9_]+", "");
-        File tmpFolder = new File(SysConstants.ROOT_PATH + dataFolder);
-        tmpFolder.mkdir();
-        if (tmpFolder.canWrite()) {
+        String dataFolder = UUID.randomUUID().toString();
+        File tmpFolder = new File(SysConstants.ROOT_PATH, dataFolder);
+        if (tmpFolder.mkdir() && tmpFolder.canWrite()) {
             writingEnabled = true;
             this.tmpFolder = SysConstants.ROOT_PATH + dataFolder + File.separator;
             tmpFolder.deleteOnExit();
